requirements.txt
langchain-core  
langchain-openai  
langchain-community  
python-dotenv
